Changed accuracy to float instead of int. Avoids accuracy = 0.

+++ main/Utils.hpp
@@ -4,6 +4,7 @@
 #include <Arduino.h>
 #include <stdint.h>
 #include <string.h>
+
 #include "eigen.h"
 
 static constexpr uint8_t STX = 0x7E;
@@ -67,8 +68,8 @@ struct GNSSData {
     float velN = 0.0f;
     float velE = 0.0f;
     float velD = 0.0f;
-    uint16_t horAcc = 0;
-    uint16_t vertAcc = 0;
+    float horAcc = 0.0f;
+    float vertAcc = 0.0f;
     uint8_t numSV = 0;
     uint8_t fixType = 0;
 };
